+++ sphinx_gallery/tests/test_backreferences.py
@@ -7,7 +7,6 @@ Testing the rst files generator
 from __future__ import division, absolute_import, print_function
 
 import sphinx_gallery.backreferences as sg
-from sphinx_gallery.utils import unicode_sample
 
 
 def test_thumbnail_div():
