--- sphinx_gallery/tests/unicode.sample
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Únicode in header
-=================
-
-Uñicode in description
-"""
-
-# Code source: Óscar Nájera
-# License: BSD 3 clause
-
-import os
-path = os.path.join('a','b')
-
-a = 'heiß'  # Unicode string
-
-import sphinx_gallery.back_references as br
-br.identify_names
-
-from sphinx_gallery.back_references import identify_names
-identify_names
